Refactor src/constants and primitives::sinsemilla::constants.
