+++ src/circuit/gadget/ecc/chip/mul/overflow.rs
@@ -1,7 +1,6 @@
-use super::Z;
+use super::{T_Q, Z};
 use crate::{
-    circuit::gadget::utilities::lookup_range_check::LookupRangeCheckConfig, constants::T_Q,
-    primitives::sinsemilla,
+    circuit::gadget::utilities::lookup_range_check::LookupRangeCheckConfig, primitives::sinsemilla,
 };
 use halo2::circuit::AssignedCell;
 use halo2::{
