+++ src/circuit/gadget/ecc/chip/mul_fixed.rs
@@ -1,9 +1,8 @@
 use super::{
     add, add_incomplete, EccBaseFieldElemFixed, EccScalarFixed, EccScalarFixedShort, FixedPoint,
-    NonIdentityEccPoint,
+    NonIdentityEccPoint, FIXED_BASE_WINDOW_SIZE, H,
 };
 use crate::circuit::gadget::utilities::decompose_running_sum::RunningSumConfig;
-use crate::constants;
 
 use std::marker::PhantomData;
 
@@ -26,15 +25,15 @@ pub mod short;
 lazy_static! {
     static ref TWO_SCALAR: pallas::Scalar = pallas::Scalar::from(2);
     // H = 2^3 (3-bit window)
-    static ref H_SCALAR: pallas::Scalar = pallas::Scalar::from(constants::H as u64);
-    static ref H_BASE: pallas::Base = pallas::Base::from(constants::H as u64);
+    static ref H_SCALAR: pallas::Scalar = pallas::Scalar::from_u64(H as u64);
+    static ref H_BASE: pallas::Base = pallas::Base::from_u64(H as u64);
 }
 
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
+#[derive(Clone, Debug, Eq, PartialEq)]
 pub struct Config<FixedPoints: super::FixedPoints<pallas::Affine>> {
-    running_sum_config: RunningSumConfig<pallas::Base, { constants::FIXED_BASE_WINDOW_SIZE }>,
+    running_sum_config: RunningSumConfig<pallas::Base, FIXED_BASE_WINDOW_SIZE>,
     // The fixed Lagrange interpolation coefficients for `x_p`.
-    lagrange_coeffs: [Column<Fixed>; constants::H],
+    lagrange_coeffs: [Column<Fixed>; H],
     // The fixed `z` for each window such that `y + z = u^2`.
     fixed_z: Column<Fixed>,
     // Decomposition of an `n-1`-bit scalar into `k`-bit windows:
@@ -57,7 +56,7 @@ impl<FixedPoints: super::FixedPoints<pallas::Affine>> Config<FixedPoints> {
     #[allow(clippy::too_many_arguments)]
     pub(super) fn configure(
         meta: &mut ConstraintSystem<pallas::Base>,
-        lagrange_coeffs: [Column<Fixed>; constants::H],
+        lagrange_coeffs: [Column<Fixed>; H],
         window: Column<Advice>,
         x_p: Column<Advice>,
         y_p: Column<Advice>,
@@ -138,7 +137,7 @@ impl<FixedPoints: super::FixedPoints<pallas::Affine>> Config<FixedPoints> {
 
             //    z_{i+1} = (z_i - a_i) / 2^3
             // => a_i = z_i - z_{i+1} * 2^3
-            let word = z_cur - z_next * pallas::Base::from(constants::H as u64);
+            let word = z_cur - z_next * pallas::Base::from(H as u64);
 
             self.coords_check(meta, q_mul_fixed_running_sum, word)
         });
@@ -156,7 +155,7 @@ impl<FixedPoints: super::FixedPoints<pallas::Affine>> Config<FixedPoints> {
         let z = meta.query_fixed(self.fixed_z, Rotation::cur());
         let u = meta.query_advice(self.u, Rotation::cur());
 
-        let window_pow: Vec<Expression<pallas::Base>> = (0..constants::H)
+        let window_pow: Vec<Expression<pallas::Base>> = (0..H)
             .map(|pow| {
                 (0..pow).fold(Expression::Constant(pallas::Base::one()), |acc, _| {
                     acc * window.clone()
@@ -228,7 +227,7 @@ impl<FixedPoints: super::FixedPoints<pallas::Affine>> Config<FixedPoints> {
             coords_check_toggle.enable(region, window + offset)?;
 
             // Assign x-coordinate Lagrange interpolation coefficients
-            for k in 0..(constants::H) {
+            for k in 0..(H) {
                 region.assign_fixed(
                     || {
                         format!(
@@ -374,12 +373,7 @@ impl<FixedPoints: super::FixedPoints<pallas::Affine>> Config<FixedPoints> {
 
         // offset_acc = \sum_{j = 0}^{NUM_WINDOWS - 2} 2^{FIXED_BASE_WINDOW_SIZE*j + 1}
         let offset_acc = (0..(NUM_WINDOWS - 1)).fold(pallas::Scalar::zero(), |acc, w| {
-            acc + (*TWO_SCALAR).pow(&[
-                constants::FIXED_BASE_WINDOW_SIZE as u64 * w as u64 + 1,
-                0,
-                0,
-                0,
-            ])
+            acc + (*TWO_SCALAR).pow(&[FIXED_BASE_WINDOW_SIZE as u64 * w as u64 + 1, 0, 0, 0])
         });
 
         // `scalar = [k * 8^84 - offset_acc]`, where `offset_acc = \sum_{j = 0}^{83} 2^{FIXED_BASE_WINDOW_SIZE*j + 1}`.
@@ -484,7 +478,7 @@ impl ScalarFixed {
             .map(|window| {
                 if let Some(window) = window {
                     let window = window.get_lower_32() as usize;
-                    assert!(window < constants::H);
+                    assert!(window < H);
                     Some(window)
                 } else {
                     None
