+++ src/constants/fixed_bases/spend_auth_g.rs
