+++ src/note/commitment.rs
@@ -6,7 +6,7 @@ use pasta_curves::pallas;
 use subtle::{ConstantTimeEq, CtOption};
 
 use crate::{
-    constants::{L_ORCHARD_BASE, NOTE_COMMITMENT_PERSONALIZATION},
+    constants::{fixed_bases::NOTE_COMMITMENT_PERSONALIZATION, L_ORCHARD_BASE},
     primitives::sinsemilla,
     spec::extract_p,
     value::NoteValue,
