+++ src/primitives/sinsemilla.rs
@@ -1,19 +1,37 @@
 //! The Sinsemilla hash function and commitment scheme.
 
-use group::Wnaf;
+use group::{Curve, Wnaf};
 use halo2::arithmetic::{CurveAffine, CurveExt};
 use pasta_curves::pallas;
 use subtle::CtOption;
 
-use crate::spec::{extract_p_bottom, i2lebsp};
-
 mod addition;
 use self::addition::IncompletePoint;
-
-mod constants;
 mod sinsemilla_s;
-pub use constants::*;
-pub(crate) use sinsemilla_s::*;
+pub use sinsemilla_s::SINSEMILLA_S;
+
+/// Number of bits of each message piece in $\mathsf{SinsemillaHashToPoint}$
+pub const K: usize = 10;
+
+/// $\frac{1}{2^K}$
+pub const INV_TWO_POW_K: [u8; 32] = [
+    1, 0, 192, 196, 160, 229, 70, 82, 221, 165, 74, 202, 85, 7, 62, 34, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+    0, 0, 0, 0, 0, 240, 63,
+];
+
+/// The largest integer such that $2^c \leq (r_P - 1) / 2$, where $r_P$ is the order
+/// of Pallas.
+pub const C: usize = 253;
+
+// Sinsemilla Q generators
+
+/// SWU hash-to-curve personalization for Sinsemilla $Q$ generators.
+pub const Q_PERSONALIZATION: &str = "z.cash:SinsemillaQ";
+
+// Sinsemilla S generators
+
+/// SWU hash-to-curve personalization for Sinsemilla $S$ generators.
+pub const S_PERSONALIZATION: &str = "z.cash:SinsemillaS";
 
 pub(crate) fn lebs2ip_k(bits: &[bool]) -> u32 {
     assert!(bits.len() == K);
@@ -22,11 +40,18 @@ pub(crate) fn lebs2ip_k(bits: &[bool]) -> u32 {
         .fold(0u32, |acc, (i, b)| acc + if *b { 1 << i } else { 0 })
 }
 
-/// The sequence of K bits in little-endian order representing an integer
-/// up to `2^K` - 1.
-pub(crate) fn i2lebsp_k(int: usize) -> [bool; K] {
-    assert!(int < (1 << K));
-    i2lebsp(int as u64)
+/// Coordinate extractor for Pallas.
+///
+/// Defined in [Zcash Protocol Spec § 5.4.9.7: Coordinate Extractor for Pallas][concreteextractorpallas].
+///
+/// [concreteextractorpallas]: https://zips.z.cash/protocol/nu5.pdf#concreteextractorpallas
+fn extract_p_bottom(point: CtOption<pallas::Point>) -> CtOption<pallas::Base> {
+    point.map(|p| {
+        p.to_affine()
+            .coordinates()
+            .map(|c| *c.x())
+            .unwrap_or_else(pallas::Base::zero)
+    })
 }
 
 /// Pads the given iterator (which MUST have length $\leq K * C$) with zero-bits to a
@@ -196,12 +221,19 @@ impl CommitDomain {
     pub(crate) fn R(&self) -> pallas::Point {
         self.R
     }
+
+    /// Returns the Sinsemilla $Q$ constant for this domain.
+    #[cfg(test)]
+    #[allow(non_snake_case)]
+    pub(crate) fn Q(&self) -> pallas::Point {
+        self.M.Q
+    }
 }
 
 #[cfg(test)]
 mod tests {
-    use super::{i2lebsp_k, lebs2ip_k, Pad, K};
-    use rand::{self, rngs::OsRng, Rng};
+    use super::{Pad, K};
+    use pasta_curves::{arithmetic::CurveExt, pallas};
 
     #[test]
     fn pad() {
@@ -242,41 +274,20 @@ mod tests {
     }
 
     #[test]
-    fn lebs2ip_k_round_trip() {
-        let mut rng = OsRng;
-        {
-            let int = rng.gen_range(0..(1 << K));
-            assert_eq!(lebs2ip_k(&i2lebsp_k(int)) as usize, int);
-        }
+    fn sinsemilla_s() {
+        use super::sinsemilla_s::SINSEMILLA_S;
+        use group::Curve;
+        use pasta_curves::arithmetic::CurveAffine;
 
-        assert_eq!(lebs2ip_k(&i2lebsp_k(0)) as usize, 0);
-        assert_eq!(lebs2ip_k(&i2lebsp_k((1 << K) - 1)) as usize, (1 << K) - 1);
-    }
+        let hasher = pallas::Point::hash_to_curve(super::S_PERSONALIZATION);
 
-    #[test]
-    fn i2lebsp_k_round_trip() {
-        {
-            let bitstring = (0..K).map(|_| rand::random()).collect::<Vec<_>>();
-            assert_eq!(
-                i2lebsp_k(lebs2ip_k(&bitstring) as usize).to_vec(),
-                bitstring
-            );
-        }
-
-        {
-            let bitstring = [false; K];
-            assert_eq!(
-                i2lebsp_k(lebs2ip_k(&bitstring) as usize).to_vec(),
-                bitstring
-            );
-        }
-
-        {
-            let bitstring = [true; K];
-            assert_eq!(
-                i2lebsp_k(lebs2ip_k(&bitstring) as usize).to_vec(),
-                bitstring
-            );
+        for j in 0..(1u32 << K) {
+            let computed = {
+                let point = hasher(&j.to_le_bytes()).to_affine().coordinates().unwrap();
+                (*point.x(), *point.y())
+            };
+            let actual = SINSEMILLA_S[j as usize];
+            assert_eq!(computed, actual);
         }
     }
 }
